Clean up code

+++ common/micdoodle8/mods/galacticraft/mars/client/render/block/GCMarsBlockRendererVine.java
@@ -53,9 +53,9 @@ public class GCMarsBlockRendererVine implements ISimpleBlockRenderingHandler
         tessellator.setBrightness(par1Block.getMixedBrightnessForBlock(var1, par2, par3, par4));
         float f = 1.0F;
         int l = par1Block.colorMultiplier(var1, par2, par3, par4);
-        float f1 = (float)(l >> 16 & 255) / 255.0F;
-        float f2 = (float)(l >> 8 & 255) / 255.0F;
-        float f3 = (float)(l & 255) / 255.0F;
+        float f1 = (l >> 16 & 255) / 255.0F;
+        float f2 = (l >> 8 & 255) / 255.0F;
+        float f3 = (l & 255) / 255.0F;
 
         if (EntityRenderer.anaglyphEnable)
         {
@@ -68,17 +68,17 @@ public class GCMarsBlockRendererVine implements ISimpleBlockRenderingHandler
         }
 
         tessellator.setColorOpaque_F(f * f1, f * f2, f * f3);
-        double d0 = (double)par2;
-        double d1 = (double)par3;
-        double d2 = (double)par4;
+        double d0 = par2;
+        double d1 = par3;
+        double d2 = par4;
 
         if (par1Block == Block.tallGrass)
         {
-            long i1 = (long)(par2 * 3129871) ^ (long)par4 * 116129781L ^ (long)par3;
+            long i1 = par2 * 3129871 ^ par4 * 116129781L ^ par3;
             i1 = i1 * i1 * 42317861L + i1 * 11L;
-            d0 += ((double)((float)(i1 >> 16 & 15L) / 15.0F) - 0.5D) * 0.5D;
-            d1 += ((double)((float)(i1 >> 20 & 15L) / 15.0F) - 1.0D) * 0.2D;
-            d2 += ((double)((float)(i1 >> 24 & 15L) / 15.0F) - 0.5D) * 0.5D;
+            d0 += ((i1 >> 16 & 15L) / 15.0F - 0.5D) * 0.5D;
+            d1 += ((i1 >> 20 & 15L) / 15.0F - 1.0D) * 0.2D;
+            d2 += ((i1 >> 24 & 15L) / 15.0F - 0.5D) * 0.5D;
         }
 
         renderBlocks.drawCrossedSquares(par1Block, var1.getBlockMetadata(par2, par3, par4), d0, d1, d2, 1.0F);
